undo bogus 32-warning avoidance
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@21765 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ mount_slash/mount_slash_int.c
@@ -2071,7 +2071,7 @@ msl_update_attributes(struct msl_fsrqinfo *q)
 	PFL_GETTIMESPEC(&ts);
 	f->fcmh_sstb.sst_mtime = ts.tv_sec;
 	f->fcmh_sstb.sst_mtime_ns = ts.tv_nsec;
-	if (q->mfsrq_off + (uint64_t)q->mfsrq_len > fcmh_2_fsz(f)) {
+	if (q->mfsrq_off + q->mfsrq_len > fcmh_2_fsz(f)) {
 		psclog_info("fid: "SLPRI_FID", "
 		    "size from %"PRId64" to %"PRId64,
 		    fcmh_2_fid(f), fcmh_2_fsz(f),
